Backend/requirements.txt
# requirements.txt - Complete Dependencies for AP FraudShield

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0

# PDF Processing
pypdf==3.17.1
pdf2image==1.16.3
pdfplumber==0.10.3

# OCR
pytesseract==0.3.10

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
piexif==1.1.3
imagehash==4.3.1

# Document Processing
python-docx==1.1.0

# Excel Processing
openpyxl==3.1.2
xlrd==2.0.1

# Machine Learning & Similarity Detection
sentence-transformers==2.2.2
transformers==4.34.1
huggingface_hub==0.18.0
torch>=1.13,<2.2
faiss-cpu==1.7.4
numpy==1.26.2
